Add files via upload
Complex species model where there are empty patches, basal species, a weak consumer and a strong consumer. The output is how many of each is left over and two graphs: one of the network, and one of the number how many of each species is in the network over time where trace 0 is empty patch, trace 1 is basal species, trace 2 is weak consumers, and trace 3 is strong consumers.